guardrails-on-eks/requirements.txt
aws-cdk-lib==2.60.0
aws-cdk.lambda-layer-kubectl-v24==2.0.78
constructs>=10.0.0,<11.0.0
